LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/return-value/navigate-expected.txt


FAIL navigate() promises assert_equals: expected object "[object NavigationHistoryEntry]" but got object "[object NavigationHistoryEntry]"
